Filter members page by active/inactive status
In order to make it easy to differentiate between active and inactive
members I have addded a dropdown that allows you to filter by active or
inactive members on the members page. This commit also allows users to
change their active status. In the future officers will be the only ones
allowed to change active status.