Move footer to bottom of page and do some basic CSS refactoring
Before this commit the footer would sometime stick to the content even
if the content didn't take up a whole page. It looks weird to have the
footer in the center of the page so this fixes that issue.

This commit also does some basic reworking of the CSS and html to make
it easier to work with.