Improve CSS rendering on smaller screens
During the meeting with the bluebonnets officers I ran into an issue
where the computer screen I was using had a much smaller aspect ratio so
the expanders on the dashboard did not show up. I have improved the CSS
to do a better job of handling that case.

There is still room for improvement to make the entire website better at
handling different sized screens, but for now it should be good enough.